+++ src/main/java/com/hanghae/final_project/domain/workspace/Schedule.java
@@ -1,5 +1,4 @@
-package com.hanghae.final_project.domain.workspace.model;
-
+package com.hanghae.final_project.domain.workspace;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -7,31 +6,25 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.sql.Clob;
 
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @Getter
 @Entity
-public class Document extends Timestamped{
+public class Schedule extends Timestamped {
 
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Id
     private Long id;
 
     @Column(nullable = false)
-    private String title;
-
-    //DATATYPE = TEXT, MIDTEXT 이런것들이 있는지 확인하고 다른 표현 방식이 있으면 수정..
-    @Lob
-    @Column(nullable = false)
     private String content;
+
+    @Column(nullable = false)
+    private String date;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "workspace_id", nullable = false)
     private WorkSpace workSpace;
-
-    private String imageUrl = null;
 }
-
