+++ src/main/java/com/hanghae/final_project/domain/workspace/document/dto/request/DocumentRequestDto.java
@@ -1,0 +1,11 @@
+package com.hanghae.final_project.domain.workspace.document.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class DocumentRequestDto {
+    private String title;
+    private String content;
+    private String imageUrl;
+
+}
