+++ src/main/java/com/hanghae/final_project/domain/workspace/document/repository/DocumentRepository.java
@@ -1,0 +1,7 @@
+package com.hanghae.final_project.domain.workspace.document.repository;
+
+import com.hanghae.final_project.domain.workspace.document.model.Document;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface DocumentRepository extends JpaRepository<Document, Long> {
+}
