--- src/main/java/com/hanghae/final_project/domain/workspace/model/Schedule.java
@@ -1,30 +1,0 @@
-package com.hanghae.final_project.domain.workspace.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-@Getter
-@Entity
-public class Schedule extends Timestamped {
-
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Id
-    private Long id;
-
-    @Column(nullable = false)
-    private String content;
-
-    @Column(nullable = false)
-    private String date;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "workspace_id", nullable = false)
-    private WorkSpace workSpace;
-}
